Updated Actuation PCB
This is the first draft layout.

It only has two of the four desired screw holes, but this is mainly due
to my refusal to make the board bigger in my tired stubbornness. There
is also no labeling on the board (other than small built in markers) as
of yet. So this will need to be added as well.